+++ src/main/java/org/kumoricon/registration/model/order/OrderService.java
@@ -5,6 +5,8 @@ import org.kumoricon.registration.model.attendee.AttendeeService;
 import org.kumoricon.registration.model.badge.BadgeService;
 import org.kumoricon.registration.model.user.User;
 import org.kumoricon.registration.model.user.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,6 +17,7 @@ public class OrderService {
     private final AttendeeService attendeeService;
     private final UserService userService;
     private final BadgeService badgeService;
+    private static final Logger log = LoggerFactory.getLogger(OrderService.class);
 
     public OrderService(OrderRepository orderRepository,
                         AttendeeService attendeeService,
@@ -40,17 +43,30 @@ public class OrderService {
     }
 
     @Transactional
-    public void addAttendeeToOrder(Integer orderId, Attendee attendee, User principal) {
+    public void saveAttendeeToOrder(Integer orderId, Attendee attendee, User principal) {
         Order order = orderRepository.findById(orderId);
-        order.addToTotalAmount(attendee.getPaidAmount());
-        orderRepository.save(order);
+        checkOrderIsNotPaid(order);
+
         attendee.setOrder(order);
         if (attendee.getBadgeNumber() == null) {
             String badgeNumber = userService.getNextBadgeNumber(principal.getUsername());
             attendee.setBadgeNumber(badgeNumber);
         }
-        attendee.setPaidAmount(badgeService.getCostForBadgeType(attendee.getBadgeId(), attendee.getAge()));
+        if (attendee.getPaidAmount() != null && principal.hasRight("attendee_override_price")) {
+            log.info("setting manual price of ${} for {}", attendee.getPaidAmount(), attendee);
+        } else {
+            attendee.setPaidAmount(badgeService.getCostForBadgeType(attendee.getBadgeId(), attendee.getAge()));
+        }
 
         attendeeService.save(attendee);
+
+        // Recalculate order total
+        order.setTotalAmount(orderRepository.getTotalByOrderId(orderId));
+        orderRepository.save(order);
+    }
+
+
+    private void checkOrderIsNotPaid(Order order) {
+        if (order.getPaid()) throw new RuntimeException("Order " + order.getId() + " is already paid and can not be changed");
     }
 }